Update MIPScode.txt
The text and main declarations are used to start out program. I used li to load 11 into v0, which tells the computer to print out characters now when the system is called. I loaded my first character '@' into a0 using la. Then i made a system call for the computer to print the character stored in a0. By subtracting 10,  I made the command in v0 equal to 1, the command to print integers. I subtracted 64 from a0 using the sub command and made it equal to 0.  I then made a system call so the integer in a0 would be printed