Update MIPScode.txt
Fixed Line 18, which wasn't indented and added Lines 22-41. 
Lines 22-24, Change the v0 register to 11, which contains the command to print out characters. When then change a0 to equal 10 which is the ASCII for a new line. I perform a system call to print out a new line
Lines 24-41, use the same method of adding and subtracting to determine which character should be printed by the system call